backend/requirements/dev.txt
# Include base requirements
-r base.txt

# Testing
pytest>=8.2,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=5.0,<6.0
factory-boy>=3.3,<4.0

# Code Quality
black>=24.4,<25.0
isort>=5.13,<6.0
flake8>=7.0,<8.0
mypy>=1.10,<2.0
django-stubs>=5.0,<6.0
djangorestframework-stubs>=3.15,<4.0

# Debugging
django-debug-toolbar>=4.4,<5.0
ipython>=8.25,<9.0